moving the file to src directory
